requirements.txt
flask
pyzmq
pymongo
docker
kubernetes